Add jaro to my kit
I'd previously written my own little URL opener script to simplify a few
things but I decided to grab jaro and use that instead.

Main use cases are:

1. Being able to define "ssh://" URLs in my keepass database so I can
   easily open remote terminals with C-U
2. Launching Teams URLs within the Firefox PWA just by clicking on them

But with the machinery in place I can now use it for generic URL
redirecting.  For example, if a site works best in Chrome I can do
that via jaro.